Docs: add index-page blurb.
The first thing I want to see when I visit a project's docs is some
information giving me an idea of how to use it, in a very rough sense.
I do *not* want to see an index with no prose, particularly if there's
only one thing in the index.

Our docs are slowly getting better, step by step.