Alert on check function exceptions.
One of the biggest concerns with alerting, monitoring, and
error-handling software is a problem in the software of the type it's
meant to prevent - who watches the watchmen?

This improves that situation a bit for Alertinator; if there is an
`Exception` thrown in one of your alert-checkers (other than an
`AlertinatorException`, which is expected), we'll treat it as a
CRITICAL-level problem, and alert you as such.

Since PHP didn't get object-orientation until late in life, it has the
notion of *errors* in addition to *Exceptions*.  We won't catch any
errors here, unless you've set up an `error_handler` to convert those
errors into (Error)Exceptions.